[BACKLOG-3857]IMetaStoreElementType base
Create base type for ElementType implementations, reduces duplication
